Introduce RunScheduler
run scheduler is responsible to execute the run ask by the user.
this allow to make sure we always only have one run executing at the same time.
